chore: Configure IMessageDialogService and IBrowser.
